Fixes to improve portability (especially to OS X, Solaris, OpenBSD).
git-svn-id: file:///home/klmitch/undernet-ircu/undernet-ircu-svn/ircu2/trunk@1091 c9e4aea6-c8fd-4c43-8297-357d70d61c8c

+++ ircd/uping.c
@@ -298,7 +298,7 @@ void uping_send(struct UPing* pptr)
   memset(buf, 0, sizeof(buf));
 
   gettimeofday(&tv, NULL);
-  sprintf(buf, " %10lu%c%6lu", tv.tv_sec, '\0', tv.tv_usec);
+  sprintf(buf, " %10lu%c%6lu", (unsigned long)tv.tv_sec, '\0', (unsigned long)tv.tv_usec);
 
   Debug((DEBUG_SEND, "send_ping: sending [%s %s] to %s.%d on %d",
 	  buf, &buf[12],
